+++ crates/integrations/connector-integration/src/connectors.rs
@@ -289,6 +289,9 @@ pub use self::flywire::Flywire;
 pub mod affirm;
 pub use self::affirm::Affirm;
 
+pub mod moneris;
+pub use self::moneris::Moneris;
+
 pub mod kount;
 pub use self::kount::Kount;
 
@@ -305,3 +308,8 @@ pub use self::boost::Boost;
 
 pub mod citigate;
 pub use self::citigate::Citigate;
+pub mod ilixium;
+pub use self::ilixium::Ilixium;
+
+pub mod worldpayraft;
+pub use self::worldpayraft::Worldpayraft;